+++ android/versioned-abis/expoview-abi43_0_0/src/main/java/abi43_0_0/expo/modules/updates/UpdatesModule.java
@@ -189,7 +189,7 @@ public class UpdatesModule extends ExportedModule {
 
       AsyncTask.execute(() -> {
         final DatabaseHolder databaseHolder = updatesService.getDatabaseHolder();
-        new RemoteLoader(getContext(), updatesService.getConfiguration(), databaseHolder.getDatabase(), updatesService.getFileDownloader(), updatesService.getDirectory())
+        new RemoteLoader(getContext(), updatesService.getConfiguration(), databaseHolder.getDatabase(), updatesService.getFileDownloader(), updatesService.getDirectory(), updatesService.getLaunchedUpdate())
           .start(
             new RemoteLoader.LoaderCallback() {
               @Override