+++ android/versioned-abis/expoview-abi44_0_0/src/main/java/abi44_0_0/expo/modules/updates/UpdatesModule.kt
@@ -190,7 +190,8 @@ class UpdatesModule(
           updatesServiceLocal.configuration,
           databaseHolder.database,
           updatesServiceLocal.fileDownloader,
-          updatesServiceLocal.directory
+          updatesServiceLocal.directory,
+          updatesServiceLocal.launchedUpdate
         )
           .start(
             object : Loader.LoaderCallback {